Fixed configuration. Now everything is working and the markers of the minimum point of the pulse and the edge threshold marker are showing on the plots

+++ analysis/include/Configuration.h
@@ -11,7 +11,7 @@ public:
 	static float GetEdgeThresholdVolts();
 	static float GetVoltMin();
 	static float GetVoltMax();
-	static float GetDigitizerResolution();
+	static int GetDigitizerResolution();
 	static float GetExpectedPulseWidthNs();
 	static float GetMinEdgeSeparationNs();
 	static float GetMaxEdgeJitterNs();
